+++ apps/participants/utils.py
@@ -45,7 +45,7 @@ def get_participant_team_id_of_user_for_a_challenge(user, challenge_id):
 
 def get_participant_team_of_user_for_a_challenge(user, challenge_id):
     """Returns the participant team object for a particular user for a particular challenge"""
-    # TODO: Remove other finctions and use this function
+    # TODO: Remove other functions and use this function
     participant_team_id = get_participant_team_id_of_user_for_a_challenge(
         user, challenge_id
     )
